+++ scripts/env.sh
@@ -1,3 +1,3 @@
 export CONTAINER_TIMEOUT="300s"
 export SERVER_STARTUP_TIME="10"
-export HL_MOUNT="/data/development-1/hyperledger"
+export HL_MOUNT="/data/development-2/hyperledger"