+++ scripts/startNetwork/setupOrg1Ca.sh
@@ -5,16 +5,8 @@ source ./scripts/env.sh
 
 header "Org1 CA"
 
-echo "Creating Org1 CA"
+msg "Starting Org1 CA"
 kubectl create -f k8s/org1/rca-org1.yaml
 
-small_sep
-
-# Wait until pod is ready
-echo "Waiting for pod"
+msg "Waiting for pod"
 kubectl wait --for=condition=ready pod -l app=rca-org1 --timeout=${CONTAINER_TIMEOUT} -n hlf
-sleep $SERVER_STARTUP_TIME
-
-small_sep
-
-kubectl exec -n hlf $(get_pods "rca-org1") -i -- bash /tmp/hyperledger/scripts/startNetwork/registerUsers/registerOrg1CaUsers.sh